refactor(scraper): extract seller profile scraper
