+++ src/xianyu/seller_profile_scraper.py
@@ -1,0 +1,115 @@
+import asyncio
+
+from playwright.async_api import Response
+
+from src.parsers import (
+    _parse_user_items_data,
+    calculate_reputation_from_ratings,
+    parse_ratings_data,
+    parse_user_head_data,
+)
+from src.utils import random_sleep
+
+
+async def scrape_user_profile(context, user_id: str) -> dict:
+    """
+    【新版】访问指定用户的个人主页，按顺序采集其摘要信息、完整的商品列表和完整的评价列表。
+    """
+    print(f"   -> 开始采集用户ID: {user_id} 的完整信息...")
+    profile_data = {}
+    page = await context.new_page()
+
+    # 为各项异步任务准备Future和数据容器
+    head_api_future = asyncio.get_event_loop().create_future()
+
+    all_items, all_ratings = [], []
+    stop_item_scrolling, stop_rating_scrolling = asyncio.Event(), asyncio.Event()
+
+    async def handle_response(response: Response):
+        # 捕获头部摘要API
+        if (
+            "mtop.idle.web.user.page.head" in response.url
+            and not head_api_future.done()
+        ):
+            try:
+                head_api_future.set_result(await response.json())
+                print(f"      [API捕获] 用户头部信息... 成功")
+            except Exception as e:
+                if not head_api_future.done():
+                    head_api_future.set_exception(e)
+
+        # 捕获商品列表API
+        elif "mtop.idle.web.xyh.item.list" in response.url:
+            try:
+                data = await response.json()
+                all_items.extend(data.get("data", {}).get("cardList", []))
+                print(f"      [API捕获] 商品列表... 当前已捕获 {len(all_items)} 件")
+                if not data.get("data", {}).get("nextPage", True):
+                    stop_item_scrolling.set()
+            except Exception as e:
+                stop_item_scrolling.set()
+
+        # 捕获评价列表API
+        elif "mtop.idle.web.trade.rate.list" in response.url:
+            try:
+                data = await response.json()
+                all_ratings.extend(data.get("data", {}).get("cardList", []))
+                print(f"      [API捕获] 评价列表... 当前已捕获 {len(all_ratings)} 条")
+                if not data.get("data", {}).get("nextPage", True):
+                    stop_rating_scrolling.set()
+            except Exception as e:
+                stop_rating_scrolling.set()
+
+    page.on("response", handle_response)
+
+    try:
+        # --- 任务1: 导航并采集头部信息 ---
+        await page.goto(
+            f"https://www.goofish.com/personal?userId={user_id}",
+            wait_until="domcontentloaded",
+            timeout=20000,
+        )
+        head_data = await asyncio.wait_for(head_api_future, timeout=15)
+        profile_data = await parse_user_head_data(head_data)
+
+        # --- 任务2: 滚动加载所有商品 (默认页面) ---
+        print("      [采集阶段] 开始采集该用户的商品列表...")
+        await random_sleep(2, 4)  # 等待第一页商品API完成
+        while not stop_item_scrolling.is_set():
+            await page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
+            try:
+                await asyncio.wait_for(stop_item_scrolling.wait(), timeout=8)
+            except asyncio.TimeoutError:
+                print("      [滚动超时] 商品列表可能已加载完毕。")
+                break
+        profile_data["卖家发布的商品列表"] = await _parse_user_items_data(all_items)
+
+        # --- 任务3: 点击并采集所有评价 ---
+        print("      [采集阶段] 开始采集该用户的评价列表...")
+        rating_tab_locator = page.locator("//div[text()='信用及评价']/ancestor::li")
+        if await rating_tab_locator.count() > 0:
+            await rating_tab_locator.click()
+            await random_sleep(3, 5)  # 等待第一页评价API完成
+
+            while not stop_rating_scrolling.is_set():
+                await page.evaluate("window.scrollTo(0, document.body.scrollHeight)")
+                try:
+                    await asyncio.wait_for(stop_rating_scrolling.wait(), timeout=8)
+                except asyncio.TimeoutError:
+                    print("      [滚动超时] 评价列表可能已加载完毕。")
+                    break
+
+            profile_data["卖家收到的评价列表"] = await parse_ratings_data(all_ratings)
+            reputation_stats = await calculate_reputation_from_ratings(all_ratings)
+            profile_data.update(reputation_stats)
+        else:
+            print("      [警告] 未找到评价选项卡，跳过评价采集。")
+
+    except Exception as e:
+        print(f"   [错误] 采集用户 {user_id} 信息时发生错误: {e}")
+    finally:
+        page.remove_listener("response", handle_response)
+        await page.close()
+        print(f"   -> 用户 {user_id} 信息采集完成。")
+
+    return profile_data
